Fixed eloquent migrations

+++ database/migrations/2015_04_27_141832_create_api_scopes_table.php
@@ -13,24 +13,26 @@ class CreateApiScopesTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('api_scopes', function (Blueprint $table) {
-			$table->bigIncrements('id');
-			$table->string('name', 512);
-			$table->string('short_description', 512);
-			$table->text('description');
-			$table->boolean('active')->default(true);
-			$table->boolean('default')->default(false);
-			$table->boolean('system')->default(false);
-			$table->timestamps();
-			// FK
-			$table->bigInteger("api_id")->unsigned()->nullable();
-			$table->index('api_id');
-			$table->foreign('api_id')
-				->references('id')
-				->on('apis')
-				->onDelete('cascade')
-				->onUpdate('no action');
-		});
+        if (!Schema::hasTable('api_scopes')) {
+            Schema::create('api_scopes', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('name', 512);
+                $table->string('short_description', 512);
+                $table->text('description');
+                $table->boolean('active')->default(true);
+                $table->boolean('default')->default(false);
+                $table->boolean('system')->default(false);
+                $table->timestamps();
+                // FK
+                $table->bigInteger("api_id")->unsigned()->nullable();
+                $table->index('api_id');
+                $table->foreign('api_id')
+                    ->references('id')
+                    ->on('apis')
+                    ->onDelete('cascade')
+                    ->onUpdate('no action');
+            });
+        }
 	}
 
 	/**
