+++ database/migrations/2015_04_27_141848_create_api_endpoints_table.php
@@ -13,45 +13,49 @@ class CreateApiEndpointsTable extends Migration
 	 */
 	public function up()
 	{
-		Schema::create('api_endpoints', function (Blueprint $table) {
-			$table->bigIncrements('id');
-			$table->boolean('active')->default(true);
-			$table->boolean('allow_cors')->default(true);
-			$table->boolean('allow_credentials')->default(true);
-			$table->text('description')->nullable();
-			$table->string('name', 255)->unique();
-			$table->timestamps();
-			$table->text("route");
-			$table->enum('http_method', array('GET', 'HEAD','POST', 'PUT', 'DELETE', 'TRACE', 'CONNECT', 'OPTIONS', 'PATCH'));
-			$table->bigInteger("rate_limit")->unsigned()->nullable();
-			//FK
-			$table->bigInteger("api_id")->unsigned();
-			$table->index('api_id');
-			$table->foreign('api_id')
-				->references('id')
-				->on('apis')
-				->onDelete('cascade')
-				->onUpdate('no action');
-		});
+        if (!Schema::hasTable('api_endpoints')) {
+            Schema::create('api_endpoints', function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->boolean('active')->default(true);
+                $table->boolean('allow_cors')->default(true);
+                $table->boolean('allow_credentials')->default(true);
+                $table->text('description')->nullable();
+                $table->string('name', 255)->unique();
+                $table->timestamps();
+                $table->text("route");
+                $table->enum('http_method', array('GET', 'HEAD', 'POST', 'PUT', 'DELETE', 'TRACE', 'CONNECT', 'OPTIONS', 'PATCH'));
+                $table->bigInteger("rate_limit")->unsigned()->nullable();
+                //FK
+                $table->bigInteger("api_id")->unsigned();
+                $table->index('api_id');
+                $table->foreign('api_id')
+                    ->references('id')
+                    ->on('apis')
+                    ->onDelete('cascade')
+                    ->onUpdate('no action');
+            });
+        }
 
-		Schema::create('endpoint_api_scopes', function ($table) {
-			$table->timestamps();
-			$table->bigInteger("api_endpoint_id")->unsigned();
-			$table->index('api_endpoint_id');
-			$table->foreign('api_endpoint_id')
-				->references('id')
-				->on('api_endpoints')
-				->onDelete('cascade')
-				->onUpdate('no action');;
-			// FK 2
-			$table->bigInteger("scope_id")->unsigned();
-			$table->index('scope_id');
-			$table->foreign('scope_id')
-				->references('id')
-				->on('api_scopes')
-				->onDelete('cascade')
-				->onUpdate('no action');
-		});
+        if (!Schema::hasTable('api_endpoints')) {
+            Schema::create('endpoint_api_scopes', function ($table) {
+                $table->timestamps();
+                $table->bigInteger("api_endpoint_id")->unsigned();
+                $table->index('api_endpoint_id');
+                $table->foreign('api_endpoint_id')
+                    ->references('id')
+                    ->on('api_endpoints')
+                    ->onDelete('cascade')
+                    ->onUpdate('no action');;
+                // FK 2
+                $table->bigInteger("scope_id")->unsigned();
+                $table->index('scope_id');
+                $table->foreign('scope_id')
+                    ->references('id')
+                    ->on('api_scopes')
+                    ->onDelete('cascade')
+                    ->onUpdate('no action');
+            });
+        }
 	}
 
 	/**
